Fixed error with renderLine for very long strings.
If asked to render a line of several thousand characters in length, the
recursive calls in the renderLine() method would appear to be entering a
stack overflow error situation, and the JVM would rightly put a rapid
stop to that.  A MUCH more efficient method for handling renderLine has
now been implemented.